[WEB] Created pages folder and moved home page to Gallery
