+++ storybook-ai/package.json
@@ -12,6 +12,7 @@
     "react-dom": "^18.2.0",
     "react-icons": "^4.12.0",
     "react-modal": "^3.16.1",
+    "react-router-dom": "^6.20.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
